Homework#22
Sorry just saving this homework#22 on here just so it doesn't get lost
