+++ src/test/java/pages/LoginPage.java
@@ -1,0 +1,31 @@
+package pages;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+
+public class LoginPage extends BasePage {
+
+
+    public LoginPage(WebDriver givenDriver) {
+        super(givenDriver);
+    }
+
+
+    public static void provideEmail(String email) {
+        WebElement emailField = driver.findElement(By.cssSelector("[type = 'email']"));
+        emailField.sendKeys(email);
+    }
+
+    public static void providePassword(String password) {
+        WebElement passwordField = driver.findElement(By.cssSelector("[type = 'password']"));
+        passwordField.sendKeys(password);
+    }
+
+    public static void clickSubmit() {
+        WebElement submitButton = driver.findElement(By.cssSelector("button[type = 'submit']"));
+        submitButton.click();
+    }
+
+
+}
